Refactor UI frame and viewmodel metadata
