+++ doc/reference/file-map.md
@@ -72,6 +72,7 @@
 ### Understand UI composition
 
 - `src/ui/ui_runtime.c`
+- `src/ui/ui_frame.c`
 - `src/ui/ui_chrome.c`
 - `src/ui/ui_inspector_panel.c`
 - `src/ui/ui_dialog_runtime.c`
